requirements.txt
Flask==0.11.1
pyDes==2.0.1
python-keystoneclient==3.9.0
python-swiftclient==3.2.0